create extra_field_entities
