+++ server/application/catalogs/commands.py
@@ -2,7 +2,7 @@ from typing import List
 
 from pydantic import Field
 
-from server.domain.catalogs.entities import ExtraField
+from server.domain.extra_fields.entities import ExtraField
 from server.domain.organizations.types import Siret
 from server.seedwork.application.commands import Command
